requirements.txt
fastapi
uvicorn[standard]
gunicorn

SQLAlchemy>=2.0
asyncpg
alembic

aiohttp

pydantic-settings

python-dotenv

psycopg2-binary
